beginning the process of separating out functions into helper file
